Move legacy left overs to graph
